Use a fixed mirror to install openssl and ca-certificates
The load balancer mirrordirector.raspbian.org might redirect to a https
host. Which leads to an error as no valid certificate is found.
Installing openssl and ca-certificates might vail if we use the
mirrordirector.raspbian.org for this as well.

So we add a fixed mirror to install these packages and then switch back.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>